Rename osmtogtfs to o2g.
