+++ o2g/ODbL-1.0.txt
