+++ o2g/__init__.py
@@ -2,4 +2,4 @@
 A simple tool to extract public transport data from OpenStreetMap.
 """
 
-__version__ = "0.4.3"
+__version__ = "0.4.4"
