+++ pyproject.toml
@@ -3,10 +3,10 @@ requires = ['flit']
 build-backend = 'flit.buildapi'
 
 [tool.flit.metadata]
-module = 'osmtogtfs'
+module = 'o2g'
 author = 'Mehdi Sadeghi'
 author-email = 'mehdi@mehdix.rog'
-home-page = 'https://github.com/hiposfer/osmtogtfs'
+home-page = 'https://github.com/hiposfer/o2g'
 requires-python = '>=3.5'
 classifiers = [
     'Development Status :: 4 - Beta',
